Tests NwbReader: Factor out compare_dicts helper function
